+++ R/qenv-eval_code.R
@@ -1,43 +1,56 @@
-#' Evaluate the code in the `qenv` environment
+#' Evaluate code in `qenv`
 #'
-#' Given code is evaluated in the `qenv` environment and appended to the `code` slot. This means
-#' that state of the environment is always a result of the stored code (if `qenv` was initialized)
-#' with reproducible code.
-#'
-#' @name eval_code
+#' @details
+#' `eval_code` evaluates given code in the `qenv` environment and appends it to the `code` slot.
+#' Thus, if the `qenv` had been instantiated empty, contents of the environment are always a result of the stored code.
 #'
 #' @param object (`qenv`)
-#' @param code (`character` or `language`) code to evaluate. Also accepts and stores comments
+#' @param code (`character` or `language`) code to evaluate. If `character`, comments are retained.
+#'
+#' @return
+#' `eval_code` returns a `qenv` object with `expr` evaluated or `qenv.error` if evaluation fails.
 #'
 #' @examples
-#' q1 <- new_qenv(env = list2env(list(a = 1)), code = quote(a <- 1))
-#' q2 <- eval_code(q1, quote(library(checkmate)))
-#' q3 <- eval_code(q2, quote(assert_number(a)))
+#' # evaluate code in qenv
+#' q <- qenv()
+#' q <- eval_code(q, "a <- 1")
+#' q <- eval_code(q, quote(library(checkmate)))
+#' q <- eval_code(q, expression(assert_number(a)))
 #'
-#' @return `qenv` object.
+#' @name eval_code
+#' @rdname qenv
+#' @aliases eval_code,qenv,character-method
+#' @aliases eval_code,qenv,language-method
+#' @aliases eval_code,qenv,expression-method
+#' @aliases eval_code,qenv.error,ANY-method
 #'
 #' @export
 setGeneric("eval_code", function(object, code) standardGeneric("eval_code"))
 
-#' @rdname eval_code
-#' @export
-setMethod("eval_code", signature = c("qenv", "expression"), function(object, code) {
-  id <- sample.int(.Machine$integer.max, size = length(code))
+setMethod("eval_code", signature = c("qenv", "character"), function(object, code) {
+  id <- sample.int(.Machine$integer.max, size = 1)
 
   object@id <- c(object@id, id)
   object@env <- rlang::env_clone(object@env, parent = parent.env(.GlobalEnv))
+  code <- paste(code, collapse = "\n")
   object@code <- c(object@code, code)
 
   current_warnings <- ""
   current_messages <- ""
 
-  for (code_line in code) {
-    # Using withCallingHandlers to capture ALL warnings and messages.
-    # Using tryCatch to capture the FIRST error and abort further evaluation.
+  parsed_code <- parse(text = code, keep.source = TRUE)
+  for (single_call in parsed_code) {
+    # Using withCallingHandlers to capture warnings and messages.
+    # Using tryCatch to capture the error and abort further evaluation.
     x <- withCallingHandlers(
       tryCatch(
         {
-          eval(code_line, envir = object@env)
+          eval(single_call, envir = object@env)
+          if (!identical(parent.env(object@env), parent.env(.GlobalEnv))) {
+            # needed to make sure that @env is always a sibling of .GlobalEnv
+            # could be changed when any new package is added to search path (through library or require call)
+            parent.env(object@env) <- parent.env(.GlobalEnv)
+          }
           NULL
         },
         error = function(e) {
@@ -45,7 +58,7 @@ setMethod("eval_code", signature = c("qenv", "expression"), function(object, cod
             message = sprintf(
               "%s \n when evaluating qenv code:\n%s",
               .ansi_strip(conditionMessage(e)),
-              paste(format_expression(code), collapse = "\n")
+              deparse1(single_call)
             ),
             class = c("qenv.error", "try-error", "simpleError"),
             trace = object@code
@@ -61,33 +74,29 @@ setMethod("eval_code", signature = c("qenv", "expression"), function(object, cod
         invokeRestart("muffleMessage")
       }
     )
+
     if (!is.null(x)) {
       return(x)
     }
-
-    object@warnings <- c(object@warnings, current_warnings)
-    object@messages <- c(object@messages, current_messages)
   }
+
+
+  object@warnings <- c(object@warnings, current_warnings)
+  object@messages <- c(object@messages, current_messages)
+
   lockEnvironment(object@env, bindings = TRUE)
   object
 })
 
-#' @rdname eval_code
-#' @export
 setMethod("eval_code", signature = c("qenv", "language"), function(object, code) {
-  code_char <- as.expression(code)
-  eval_code(object, code_char)
+  eval_code(object, code = paste(lang2calls(code), collapse = "\n"))
 })
 
-#' @rdname eval_code
-#' @export
-setMethod("eval_code", signature = c("qenv", "character"), function(object, code) {
-  eval_code(object, code = parse(text = code, keep.source = FALSE))
+setMethod("eval_code", signature = c("qenv", "expression"), function(object, code) {
+  eval_code(object, code = paste(lang2calls(code), collapse = "\n"))
 })
 
-#' @rdname eval_code
-#' @export
-setMethod("eval_code", signature = "qenv.error", function(object, code) {
+setMethod("eval_code", signature = c("qenv.error", "ANY"), function(object, code) {
   object
 })