+++ R/qenv-get_code.R
@@ -1,17 +1,27 @@
 #' Get code from `qenv`
 #'
-#' @name get_code
+#' @details
+#' `get_code` retrieves the code stored in the `qenv`. `...` passes arguments to methods.
+#'
 #' @param object (`qenv`)
-#' @param deparse (`logical(1)`) if the returned code should be converted to character.
-#' @return named `character` with the reproducible code.
+#' @param deparse (`logical(1)`) flag specifying whether to return code as `character` or `expression`.
+#' @param ... see `Details`
+#'
+#' @return
+#' `get_code` returns the traced code (from `@code` slot) in the form specified by `deparse`.
+#'
 #' @examples
-#' q1 <- new_qenv(env = list2env(list(a = 1)), code = quote(a <- 1))
-#' q2 <- eval_code(q1, code = quote(b <- a))
-#' q3 <- eval_code(q2, code = quote(d <- 2))
-#' get_code(q3)
-#' get_code(q3, deparse = FALSE)
+#' # retrieve code
+#' get_code(q)
+#' get_code(q, deparse = FALSE)
+#'
+#' @name get_code
+#' @rdname qenv
+#' @aliases get_code,qenv-method
+#' @aliases get_code,qenv.error-method
+#'
 #' @export
-setGeneric("get_code", function(object, deparse = TRUE) {
+setGeneric("get_code", function(object, deparse = TRUE, ...) {
   # this line forces evaluation of object before passing to the generic
   # needed for error handling to work properly
   grDevices::pdf(nullfile())
@@ -21,26 +31,26 @@ setGeneric("get_code", function(object, deparse = TRUE) {
   standardGeneric("get_code")
 })
 
-#' @rdname get_code
-#' @export
 setMethod("get_code", signature = "qenv", function(object, deparse = TRUE) {
   checkmate::assert_flag(deparse)
   if (deparse) {
-    format_expression(object@code)
+    if (length(object@code) == 0) {
+      object@code
+    } else {
+      paste(object@code, collapse = "\n")
+    }
   } else {
-    object@code
+    parse(text = paste(c("{", object@code, "}"), collapse = "\n"), keep.source = TRUE)
   }
 })
 
-#' @rdname get_code
-#' @export
-setMethod("get_code", signature = "qenv.error", function(object) {
+setMethod("get_code", signature = "qenv.error", function(object, ...) {
   stop(
     errorCondition(
       sprintf(
         "%s\n\ntrace: \n %s\n",
         conditionMessage(object),
-        paste(format_expression(object$trace), collapse = "\n ")
+        paste(object$trace, collapse = "\n ")
       ),
       class = c("validation", "try-error", "simpleError")
     )